Removed incrdumpswget.py
incrdumpswget.py was a script that supported the archiving of the
incremental dumps by downloading the dumps using wget. However,
the script has been lacking in updates for quite a while, and is not
even used to do the real archiving.

This script is hence deleted to avoid third-party users from using one
that is not properly supported.